Rename web100 package.
